% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_credentials}
\alias{check_credentials}
\title{Check if the provided credentials are valid.}
\usage{
check_credentials(credentials)
}
\arguments{
\item{credentials}{A list or data frame containing the credentials to be checked.}
}
\value{
TRUE if the credentials are valid, an error otherwise
}
\description{
This function checks if the provided credentials contain a "username" and "password" field,
and if both fields are of type character. It also verifies that the credentials object has a length of 2.
}
\examples{
\dontrun{
credentials <- list(username = "john_doe", password = "password123")
check_credentials(credentials)
#> [1] TRUE
}
}
