% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etnservice-package.R
\docType{package}
\name{etnservice-package}
\alias{etnservice}
\alias{etnservice-package}
\title{etnservice: Serve Data from the European Tracking Network}
\description{
Provides API endpoints to the European Tracking Network. Designed to be used with OpenCPU and the 'etn' package.
}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})
}

Other contributors:
\itemize{
  \item Research Institute for Nature and Forest (INBO) (https://www.vlaanderen.be/inbo/en-gb/) [copyright holder]
  \item LifeWatch Belgium (https://lifewatch.be) [funder]
}

}
\keyword{internal}
