% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_deployment_logs.R
\name{get_acoustic_deployment_logs}
\alias{get_acoustic_deployment_logs}
\title{Retrieve log files/diagnostic information for deployments of acoustic
receivers.}
\usage{
get_acoustic_deployment_logs(
  credentials = list(username = Sys.getenv("ETN_USER"), password = Sys.getenv("ETN_PWD")),
  deployment_id,
  limit = FALSE
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}

\item{deployment_id}{Integer (vector). One or more deployment identifiers.}

\item{limit}{Logical. Limit the number of returned records to 100 (useful
for testing purposes). Defaults to \code{FALSE}.}
}
\value{
A tibble with receiver diagnostics data. The data is embedded in a
\code{log_data} column that is JSON formatted.
}
\description{
This function retrieves log files and diagnostic information for deployments
of acoustic receivers from the ETN database. The returned data includes
deployment ID, receiver ID, station name, datetime, record type, and log data
in JSON format.
}
\examples{
get_acoustic_deployment_logs(deployment_id = 6028)

get_acoustic_deployment_logs(deployment_id = c(53790, 1758), limit = TRUE)
}
