% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_detections_page.R
\name{get_acoustic_detections_page}
\alias{get_acoustic_detections_page}
\title{Getting a single page of a multi page acoustic detections query}
\usage{
get_acoustic_detections_page(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd")),
  next_id_pk = 0,
  page_size = 1e+06,
  start_date = NULL,
  end_date = NULL,
  acoustic_tag_id = NULL,
  animal_project_code = NULL,
  scientific_name = NULL,
  acoustic_project_code = NULL,
  receiver_id = NULL,
  station_name = NULL,
  count = FALSE
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the
ETN database.}

\item{next_id_pk}{The next primary key to fetch. All detections have a
sequential id, this key allows us to read the view top to bottom, but
filter out any records before the one we've already fetched. By default,
start reading at the first detection.}

\item{page_size}{The number of records to retrieve.}

\item{start_date}{Character. Start date (inclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{end_date}{Character. End date (exclusive) in ISO 8601 format (
\code{yyyy-mm-dd}, \code{yyyy-mm} or \code{yyyy}).}

\item{acoustic_tag_id}{Character (vector). One or more acoustic tag ids.}

\item{animal_project_code}{Character (vector). One or more animal project
codes. Case-insensitive.}

\item{scientific_name}{Character (vector). One or more scientific names.}

\item{acoustic_project_code}{Character (vector). One or more acoustic project
codes. Case-insensitive.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{station_name}{Character (vector). One or more deployment station names.}

\item{count}{Logical. If set to \code{TRUE} a data.frame is returned with a single
column: the count of the number of records returned by the query.
\code{page_size} is ignored.}
}
\value{
A tibble with acoustic detections data, with length \code{page_size} or
smaller. Including a column with the primary key of the next detection.
}
\description{
A view was implemented that returns the next primary key id, this allows for
optimized page wise detections querying to the database as no offset needs to
be included.
}
