% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acoustic_receivers.R
\name{get_acoustic_receivers}
\alias{get_acoustic_receivers}
\title{Get acoustic receiver data}
\usage{
get_acoustic_receivers(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd")),
  receiver_id = NULL,
  status = NULL
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}

\item{receiver_id}{Character (vector). One or more receiver identifiers.}

\item{status}{Character. One or more statuses, e.g. \code{available} or \code{broken}.}
}
\value{
A tibble with acoustic receiver data, sorted by \code{receiver_id}. See
also
\href{https://inbo.github.io/etn/articles/etn_fields.html}{field definitions}.
Values for \code{owner_organization} will only be visible if you are member of
the group.
}
\description{
Get data for acoustic receivers, with options to filter results.
}
\examples{
# Set credentials
credentials <- list(
   username = Sys.getenv("userid"),
   password = Sys.getenv("pwd")
 )

# Get all acoustic receivers
get_acoustic_receivers(credentials)

# Get lost and broken acoustic receivers
get_acoustic_receivers(credentials, status = c("lost", "broken"))

# Get a specific acoustic receiver
get_acoustic_receivers(credentials, receiver_id = "VR2W-124070")
}
