% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{Version information about etnservice}
\usage{
get_version()
}
\value{
A list with the following elements:
\item{ls}{A character vector of the objects in the environment.}
\item{version}{The version of the etnservice package.}
}
\description{
Users of the etn package might (probably will) have a local version of
etnservice. If there are differences between the version deployed on the API
and the version installed locally, this might lead to inconsistent results
when switching between local database queries and queries over the API. This
function returns version information of the installed version of etnservice,
when called over the api, it'll return version information on the version of
etnservice deployed via opencpu.
}
\examples{
get_version()
}
