% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_ids.R
\name{list_animal_ids}
\alias{list_animal_ids}
\title{List all available animal ids}
\usage{
list_animal_ids(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{id_pk} present in \code{common.animal_release}.
}
\description{
List all available animal ids
}
