% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_animal_project_codes.R
\name{list_animal_project_codes}
\alias{list_animal_project_codes}
\title{List all available animal project codes}
\usage{
list_animal_project_codes(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{project_code} of \code{type = "animal"} in
\code{project.sql}.
}
\description{
List all available animal project codes
}
