% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deployment_ids.R
\name{list_deployment_ids}
\alias{list_deployment_ids}
\title{List all available receiver ids}
\usage{
list_deployment_ids(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{id_pk} present in \code{acoustic.deployments}.
}
\description{
List all available receiver ids
}
