% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_receiver_ids.R
\name{list_receiver_ids}
\alias{list_receiver_ids}
\title{List all available receiver ids}
\usage{
list_receiver_ids(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{receiver} present in \code{acoustic.receivers}.
}
\description{
List all available receiver ids
}
