% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_station_names.R
\name{list_station_names}
\alias{list_station_names}
\title{List all available station names}
\usage{
list_station_names(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{station_name} present in
\code{acoustic.deployments}.
}
\description{
List all available station names
}
