% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tag_serial_numbers.R
\name{list_tag_serial_numbers}
\alias{list_tag_serial_numbers}
\title{List all available tag serial numbers}
\usage{
list_tag_serial_numbers(
  credentials = list(username = Sys.getenv("userid"), password = Sys.getenv("pwd"))
)
}
\arguments{
\item{credentials}{A list with the username and password to connect to the ETN database.}
}
\value{
A vector of all unique \code{tag_serial_numbers} present in
\code{common.tag_device}.
}
\description{
List all available tag serial numbers
}
