% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Return the installed version of etnservice}
\usage{
version()
}
\value{
A character vector with the version number of the installed version
of etnservice.
}
\description{
This function is useful to check what version of etnservice is deployed to
openCPU. This way you can validate that any local version of etnservice is
the same as the one you are adressing via the API.
}
\details{
If you have multiple versions of etnservice installed, multiple version
numbers will be returned.
}
\examples{
version()
}
